Moved Eiten to a separate file
